Move files from vostok.zookeeper.client Add AddAuthenticationInfo in IZooKeeperClient
